chore: factor out graphql logic from trino logic
